python/requirements.txt
grpcio>=1.60.0
grpcio-tools>=1.60.0
protobuf>=4.25.0
pytest>=7.0
pytest-asyncio>=0.23.0
